支持token写入redis缓存

+++ src/main/java/com/iware/lottery/Constants.java
@@ -24,7 +24,7 @@ public final class Constants {
     /**
      * token有效期（小时）
      */
-    public static final int TOKEN_EXPIRES_HOUR = 72;
+    public static final long TOKEN_EXPIRES_HOUR = 72;
 
     /**
      * 存放Authorization的header字段
